+++ st2actions/tests/unit/test_mistral_v2.py
@@ -14,18 +14,20 @@
 # limitations under the License.
 
 import copy
-import json
 import uuid
 
 import mock
+from mock import call
 import requests
 import six
 import yaml
 
 from mistralclient.api.base import APIException
+from mistralclient.api.v2 import action_executions
+from mistralclient.api.v2 import executions
 from mistralclient.api.v2 import workbooks
 from mistralclient.api.v2 import workflows
-from mistralclient.api.v2 import executions
+from oslo_config import cfg
 
 # XXX: actionsensor import depends on config being setup.
 import st2tests.config as tests_config
@@ -36,18 +38,14 @@ from st2actions.handlers.mistral import MistralCallbackHandler
 from st2actions.runners.localrunner import LocalShellRunner
 from st2actions.runners.mistral.v2 import MistralRunner
 from st2common.constants import action as action_constants
-from st2common.models.api.auth import TokenAPI
 from st2common.models.api.action import ActionAPI
 from st2common.models.api.notification import NotificationsHelper
 from st2common.models.db.liveaction import LiveActionDB
 from st2common.persistence.action import Action
 from st2common.persistence.liveaction import LiveAction
-from st2common.services import access as access_service
 from st2common.services import action as action_service
 from st2common.transport.liveaction import LiveActionPublisher
 from st2common.transport.publishers import CUDPublisher
-from st2common.util import isotime
-from st2common.util import date as date_utils
 from st2tests import DbTestCase
 from st2tests import http
 from st2tests.fixturesloader import FixturesLoader
@@ -133,14 +131,7 @@ WF2_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
 WF2_EXEC['workflow_name'] = WF2_NAME
 
 # Action executions requirements
-ACTION_CONTEXT = {'user': 'stanley'}
 ACTION_PARAMS = {'friend': 'Rocky'}
-
-# Token for auth test cases
-TOKEN_API = TokenAPI(
-    user=ACTION_CONTEXT['user'], token=uuid.uuid4().hex,
-    expiry=isotime.format(date_utils.get_datetime_utc_now(), offset=False))
-TOKEN_DB = TokenAPI.to_model(TOKEN_API)
 
 NON_EMPTY_RESULT = 'non-empty'
 
@@ -153,16 +144,21 @@ NON_EMPTY_RESULT = 'non-empty'
                    mock.MagicMock(side_effect=MockLiveActionPublisher.publish_create))
 @mock.patch.object(LiveActionPublisher, 'publish_state',
                    mock.MagicMock(side_effect=MockLiveActionPublisher.publish_state))
-class TestMistralRunner(DbTestCase):
+class MistralRunnerTest(DbTestCase):
 
     @classmethod
     def setUpClass(cls):
-        super(TestMistralRunner, cls).setUpClass()
+        super(MistralRunnerTest, cls).setUpClass()
         runners_registrar.register_runner_types()
 
         for _, fixture in six.iteritems(FIXTURES['actions']):
             instance = ActionAPI(**fixture)
             Action.add_or_update(ActionAPI.to_model(instance))
+
+    def tearDown(self):
+        super(MistralRunnerTest, self).tearDown()
+        cfg.CONF.set_default('max_attempts', 2, group='mistral')
+        cfg.CONF.set_default('retry_wait', 1, group='mistral')
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -199,58 +195,6 @@ class TestMistralRunner(DbTestCase):
                     'st2_context': {
                         'endpoint': 'http://0.0.0.0:9101/v1/actionexecutions',
                         'parent': {
-                            'execution_id': str(execution.id)
-                        },
-                        'notify': {},
-                        'skip_notify_tasks': []
-                    }
-                }
-            }
-        }
-
-        executions.ExecutionManager.create.assert_called_with(
-            WF1_NAME, workflow_input=workflow_input, env=env)
-
-    @mock.patch.object(
-        workflows.WorkflowManager, 'list',
-        mock.MagicMock(return_value=[]))
-    @mock.patch.object(
-        workflows.WorkflowManager, 'get',
-        mock.MagicMock(return_value=WF1))
-    @mock.patch.object(
-        workflows.WorkflowManager, 'create',
-        mock.MagicMock(return_value=[WF1]))
-    @mock.patch.object(
-        executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
-    @mock.patch.object(
-        access_service, 'create_token',
-        mock.MagicMock(return_value=TOKEN_DB))
-    def test_launch_workflow_with_auth(self):
-        MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
-        liveaction = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS, context=ACTION_CONTEXT)
-        liveaction, execution = action_service.request(liveaction)
-        liveaction = LiveAction.get_by_id(str(liveaction.id))
-        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_RUNNING)
-
-        mistral_context = liveaction.context.get('mistral', None)
-        self.assertIsNotNone(mistral_context)
-        self.assertEqual(mistral_context['execution_id'], WF1_EXEC.get('id'))
-        self.assertEqual(mistral_context['workflow_name'], WF1_EXEC.get('workflow_name'))
-
-        workflow_input = copy.deepcopy(ACTION_PARAMS)
-        workflow_input.update({'count': '3'})
-
-        env = {
-            'st2_execution_id': str(execution.id),
-            'st2_liveaction_id': str(liveaction.id),
-            '__actions': {
-                'st2.action': {
-                    'st2_context': {
-                        'auth_token': TOKEN_DB.token,
-                        'endpoint': 'http://0.0.0.0:9101/v1/actionexecutions',
-                        'parent': {
-                            'user': liveaction.context['user'],
                             'execution_id': str(execution.id)
                         },
                         'notify': {},
@@ -640,8 +584,8 @@ class TestMistralRunner(DbTestCase):
                                         '["a", "b", "c"]')
 
     @mock.patch.object(
-        requests, 'request',
-        mock.MagicMock(return_value=http.FakeResponse({}, 200, 'OK')))
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(return_value=None))
     def test_callback(self):
         liveaction = LiveActionDB(
             action='core.local', parameters={'cmd': 'uname -a'},
@@ -654,10 +598,62 @@ class TestMistralRunner(DbTestCase):
         liveaction, execution = action_service.request(liveaction)
         liveaction = LiveAction.get_by_id(str(liveaction.id))
         self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_SUCCEEDED)
-        requests.request.assert_called_with('PUT', liveaction.callback['url'],
-                                            data=json.dumps({'state': 'SUCCESS',
-                                                             'output': NON_EMPTY_RESULT}),
-                                            headers={'content-type': 'application/json'})
+        action_executions.ActionExecutionManager.update.assert_called_with(
+            '12345', state='SUCCESS', output=NON_EMPTY_RESULT)
+
+    @mock.patch.object(
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(side_effect=[
+            requests.exceptions.ConnectionError(),
+            None]))
+    def test_callback_retry(self):
+        cfg.CONF.set_default('max_attempts', 3, group='mistral')
+        cfg.CONF.set_default('retry_wait', 0, group='mistral')
+
+        liveaction = LiveActionDB(
+            action='core.local', parameters={'cmd': 'uname -a'},
+            callback={
+                'source': 'mistral',
+                'url': 'http://localhost:8989/v2/action_executions/12345'
+            }
+        )
+
+        liveaction, execution = action_service.request(liveaction)
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_SUCCEEDED)
+
+        calls = [call('12345', state='SUCCESS', output=NON_EMPTY_RESULT) for i in range(0, 2)]
+        action_executions.ActionExecutionManager.update.assert_has_calls(calls)
+
+    @mock.patch.object(
+        action_executions.ActionExecutionManager, 'update',
+        mock.MagicMock(side_effect=[
+            requests.exceptions.ConnectionError(),
+            requests.exceptions.ConnectionError(),
+            requests.exceptions.ConnectionError(),
+            requests.exceptions.ConnectionError(),
+            None]))
+    def test_callback_retry_exhausted(self):
+        cfg.CONF.set_default('max_attempts', 3, group='mistral')
+        cfg.CONF.set_default('retry_wait', 0, group='mistral')
+
+        liveaction = LiveActionDB(
+            action='core.local', parameters={'cmd': 'uname -a'},
+            callback={
+                'source': 'mistral',
+                'url': 'http://localhost:8989/v2/action_executions/12345'
+            }
+        )
+
+        liveaction, execution = action_service.request(liveaction)
+        liveaction = LiveAction.get_by_id(str(liveaction.id))
+        self.assertEqual(liveaction.status, action_constants.LIVEACTION_STATUS_SUCCEEDED)
+
+        # This test initially setup mock for action_executions.ActionExecutionManager.update
+        # to fail the first 4 times and return success on the 5th times. The max attempts
+        # is set to 3. We expect only 3 calls to pass thru the update method.
+        calls = [call('12345', state='SUCCESS', output=NON_EMPTY_RESULT) for i in range(0, 3)]
+        action_executions.ActionExecutionManager.update.assert_has_calls(calls)
 
     def test_build_context(self):
         parent = {