+++ st2common/st2common/models/api/action.py
@@ -13,6 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import copy
+
 from st2common.util import isotime
 from st2common.util import schema as util_schema
 from st2common import log as logging
@@ -28,9 +30,12 @@ from st2common.constants.action import LIVEACTION_STATUSES
 from st2common.models.system.common import ResourceReference
 
 
-__all__ = ['ActionAPI',
-           'LiveActionAPI',
-           'RunnerTypeAPI']
+__all__ = [
+    'ActionAPI',
+    'ActionCreateAPI',
+    'LiveActionAPI',
+    'RunnerTypeAPI'
+]
 
 
 LOG = logging.getLogger(__name__)
@@ -119,7 +124,9 @@ class RunnerTypeAPI(BaseAPI):
 
 
 class ActionAPI(BaseAPI):
-    """The system entity that represents a Stack Action/Automation in the system."""
+    """
+    The system entity that represents a Stack Action/Automation in the system.
+    """
 
     model = ActionDB
     schema = {
@@ -233,6 +240,32 @@ class ActionAPI(BaseAPI):
                           ref=ref)
 
         return model
+
+
+class ActionCreateAPI(ActionAPI):
+    """
+    API model for create action operations.
+    """
+    schema = copy.deepcopy(ActionAPI.schema)
+    schema['properties']['data_files'] = {
+        'description': 'Optional action script and data files which are written to the filesystem.',
+        'type': 'array',
+        'items': {
+            'type': 'object',
+            'properties': {
+                'file_path': {
+                    'type': 'string',
+                    'required': True
+                },
+                'content': {
+                    'type': 'string',
+                    'required': True
+                },
+            },
+            'additionalProperties': False
+        },
+        'default': {}
+    }
 
 
 class LiveActionAPI(BaseAPI):