+++ st2common/st2common/util/system_info.py
@@ -16,10 +16,22 @@
 import os
 import socket
 
+__all__ = [
+    'get_host_info',
+    'get_process_info'
+]
+
+
+def get_host_info():
+    host_info = {
+        'hostname': socket.gethostname()
+    }
+    return host_info
+
 
 def get_process_info():
-    runner_info = {
+    process_info = {
         'hostname': socket.gethostname(),
         'pid': os.getpid()
     }
-    return runner_info
+    return process_info